+++ src/auth/mod.rs
@@ -2,6 +2,8 @@
 
 #[cfg(feature = "approle")]
 pub mod approle;
+#[cfg(feature = "cert-auth")]
+pub mod cert;
 #[cfg(feature = "kubernetes-auth")]
 pub mod kubernetes;
 #[cfg(feature = "token")]